Show locality places in selected objects list
Major improvements to locality visualization:

1. Selected Objects List Enhancement:
   - Now shows transition followed by all locality places
   - Input places: ↓ P1 (5 tokens) - lighter color, indented
   - Output places: ↑ P2 (0 tokens) - darker color, indented
   - Visual hierarchy: transition at top, places nested below
   - Color coordination: places derive from transition color

2. Float Format Fix:
   - Changed token_balance format from :+d to :+.0f
   - Changed total_weight format to :.0f
   - Updated return types: int -> float
   - Handles arc weights as floats (continuous Petri nets)

3. Waiting State Display:
   - Shows locality structure when simulation not running
   - Lists all transitions with their locality info
   - Clear 'Start simulation' call-to-action

Example display:
  [Red] T1 [CON] (T1)                    [✕]
      [Light Red] ↓ P1 (5 tokens)
      [Light Red] ↓ P2 (3 tokens)
      [Dark Red] ↑ P3 (0 tokens)

Benefits:
- See complete locality at a glance
- Know what will be plotted
- Token counts visible
- Color shows relationship to transition

Fixes ValueError in transition properties dialog.

+++ src/shypn/diagnostic/locality_analyzer.py
@@ -136,7 +136,7 @@ class LocalityAnalyzer:
         """
         return sum(place.tokens for place in locality.output_places)
     
-    def _calculate_token_balance(self, locality: Locality) -> int:
+    def _calculate_token_balance(self, locality: Locality) -> float:
         """Calculate token balance (inputs - outputs).
         
         A positive balance indicates more tokens in input places,
@@ -148,9 +148,9 @@ class LocalityAnalyzer:
         Returns:
             Token balance (input_tokens - output_tokens)
         """
-        return self._count_input_tokens(locality) - self._count_output_tokens(locality)
+        return float(self._count_input_tokens(locality) - self._count_output_tokens(locality))
     
-    def _calculate_total_weight(self, locality: Locality) -> int:
+    def _calculate_total_weight(self, locality: Locality) -> float:
         """Calculate total arc weight.
         
         Sums the weights of all arcs (input and output) in the locality.
@@ -161,9 +161,9 @@ class LocalityAnalyzer:
         Returns:
             Sum of all arc weights
         """
-        total = 0
+        total = 0.0
         for arc in locality.input_arcs + locality.output_arcs:
-            total += getattr(arc, 'weight', 1)
+            total += float(getattr(arc, 'weight', 1))
         return total
     
     def _check_firing_potential(self, locality: Locality) -> bool:
